requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Flask==3.1.3
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
playwright==1.58.0
playwright-stealth==2.0.3
pyee==13.0.1
python-dotenv==1.2.2
requests==2.33.1
schedule==1.2.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
